refactor(gzip): move to another folder
