+++ internal/compressor/gzip/compressor.go
